pyproject.toml
[project]
name = "vision-detector"
version = "0.1.0"
description = "Détecteur temps réel : émotions, gestes, posture et présence via webcam"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mediapipe>=0.10.0",
    "opencv-python>=4.8.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "numpy>=1.24",
    "pyyaml>=6.0",
    "hsemotion-onnx>=0.2.0",
    "onnxruntime>=1.16.0",
    "timm>=0.9.0",
    "customtkinter>=5.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
